requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
frozenlist==1.7.0
idna==3.10
iso8601==2.1.0
multidict==6.5.0
propcache==0.3.2
python-dotenv==1.1.0
python-vlc==3.0.21203
requests==2.32.4
soupsieve==2.7
twitchio==2.10.0
typing_extensions==4.14.0
urllib3==2.5.0
yarl==1.20.1
yt-dlp==2025.6.9
